requirements.txt
rumps
watchdog
requests
openai
keyring
pyinstaller
pillow
pillow-heif
